Add UIView extension for dark appearance
